% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_with_data.R
\name{sim_clust_obs}
\alias{sim_clust_obs}
\title{Simulate microbiome sample data through cluster model}
\usage{
sim_clust_obs(
  prob_matrix,
  initial_state,
  n_clust,
  n_person,
  sigma1,
  n_species,
  n_depth,
  obs
)
}
\arguments{
\item{n_species}{number of species}

\item{n_depth}{sequence depth}

\item{n_samples}{number of samples}

\item{Sigma}{correlation matrix for muti-nomal distribution}
}
\value{
the simulated data
}
\description{
Simulate microbiome abundance count data with user-specified number of samples, species and sequencing depth
}
\examples{
sim_clust_obs(initial_state = 1, n_clust =4, n_person = 1, obs=preg_data)
}
