% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_without_data.R
\name{sim_ts_cluster}
\alias{sim_ts_cluster}
\title{Simulate microbiome clustered data}
\usage{
sim_ts_cluster(
  prob_matrix,
  initial_state,
  n_steps,
  sigma1,
  sigma2,
  n_species,
  n_depth
)
}
\arguments{
\item{prob_matrix}{transition matrix of markov chain}

\item{initial_state}{the initial state of the markov chain}

\item{n_steps}{number of time points}

\item{n_species}{number of species}

\item{n_depth}{sequence depth}

\item{Sigma1}{correlation matrix for centers}

\item{Sigma2}{correlation matrix for the variation around centers}
}
\value{
the simulated data
}
\description{
To simulate microbiome longitudinal count data with clustering and markov chain based on user input transition matrix
}
\examples{
sim_ts_cluster(prob_matrix=matrix(c(0.70, 0.2,0.1,0.3,0.4,0.3,0,0.4,0.6), nrow=3, byrow=TRUE), n_steps=25, initial_state = 3, sigma1=0.1, sigma2=0.01, n_species=100, n_depth=1000)
}
